# data/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.1.1"
  title: Limei Dragonfly Data Api
# during dev, should point to your local machine
host: 127.0.0.1:8181
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  token:
    type: apiKey
    name: token
    in: header
paths:
  /auth:
    x-swagger-router-controller: auth
    get:
      description: "获取验证信息"
      operationId: getAuth
      security:
        - token: []
      parameters:
        - name: token
          in: query
          description: "token值"
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AuthModel"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    post:
      description: "注册一个新的验证token"
      operationId: postAuth
      security:
        - token: []
      parameters:
        - name: email
          in: query
          description: "新token绑定的email"
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AuthModel"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    put:
      description: "更新token绑定信息"
      operationId: putAuth
      security:
        - token: []
      parameters:
        - name: token
          in: query
          description: "token值"
          required: true
          type: string
        - name: email
          in: formData
          description: "关联email"
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AuthModel"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    delete:
      description: "删除token"
      operationId: deleteAuth
      security:
        - token: []
      parameters:
        - name: token
          in: query
          description: "token值"
          required: true
          type: string
      responses:
        "200":
          description: Success
...
definitions:
  AuthModel:
    properties:
      token:
        type: string
        description: "用于用户访问时的验证Token"
      email:
        type: string
        description: "关联该token的Email，用于自动发送相关邮件"
      state:
        type: boolean
        description: "当前状态：true为允许使用，false为不允许使用"
...
  Error:
    properties:
      status:
        type: number
        description: "错误状态码，http status"
      message:
        type: string
        description: "错误的描述信息，和具体业务相关"